refactor: InputValidator 도메인으로 분류 변경
도메인과 UI를 동시에 사용하는 함수를 App.js에 새로 구현 후 사용
